[tests][refactor] Refactor and tests
