+++ src/manager/agents/common-module/module/keynodes/ScComponentManagerKeynodes.cpp
