+++ src/manager/agents/common-module/module/keynodes/ScComponentManagerKeynodes.hpp
