+++ src/manager/agents/common-module/module/utils/common_utils.hpp
