+++ src/manager/agents/common-module/utils/sc_component_utils.hpp
