+++ src/manager/agents/install-module/CMakeLists.txt
@@ -5,7 +5,7 @@ include_directories(${CMAKE_CURRENT_LIST_DIR} ${SC_MEMORY_SRC} ${SC_KPM_SRC} ${S
 add_library(install-lib SHARED ${SOURCES})
 target_link_libraries(install-lib sc-memory sc-agents-common common-lib)
 
-sc_codegen_ex(install-lib ${CMAKE_CURRENT_LIST_DIR} ${CMAKE_CURRENT_LIST_DIR}/generated)
+sc_codegen_ex(install-lib ${CMAKE_CURRENT_LIST_DIR}/module ${CMAKE_CURRENT_LIST_DIR}/generated)
 
 if(${SC_CLANG_FORMAT_CODE})
   target_clangformat_setup(install-lib)
