+++ src/manager/agents/search-module/utils/sc_component_manager_command_search.cpp
@@ -7,9 +7,8 @@
 #include <algorithm>
 
 #include "sc_component_manager_command_search.hpp"
-#include "../agents/sc_component_manager_agent_search.hpp"
-#include "../../common-module/keynodes/ScComponentManagerKeynodes.hpp"
-#include "../../common-module/utils/common_utils.hpp"
+#include "../../common-module/module/keynodes/ScComponentManagerKeynodes.hpp"
+#include "../../common-module/module/utils/common_utils.hpp"
 
 ScAddrVector ScComponentManagerCommandSearch::Execute(ScMemoryContext * context, ScAddr const & actionAddr)
 {
